inference_/requirements.txt
numpy==1.23.3
opencv_python_headless==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
scikit_image==0.19.3
skimage==0.0
tensorflow==2.10.0
tensorflow_gpu==2.10.0
torch==1.12.1
tqdm==4.64.1
